refactor(audit): ARC18 F008 isolate skin diagnostics
