+++ Editor/MainWindow.h
@@ -46,6 +46,7 @@ class MainWindow;
 class QLabel;
 class TitleBar;
 class UpdateToast;
+namespace SkinSwitchStorm { void run(MainWindow& window); }
 
 // MainWindow's implementation is split across several translation units (all
 // listed in Editor.pro SOURCES). When looking for a method, check the matching
@@ -68,10 +69,6 @@ public:
 	~MainWindow();
 	void doChecks();
 	void runDeviceSelector();
-	// Diagnostic behind --skin-switch-storm: drives the real skin/dark menu
-	// actions through a revisit-heavy sequence on the live window and reports
-	// the main toolbar's health after every switch, then exits the process.
-	void startSkinSwitchStorm();
 	void load(QString path);
 	void save(FilterTable* filterTable, QString path);
 	bool isEmpty();
@@ -129,6 +126,8 @@ private slots:
 	void on_actionResetAllFileSpecificPreferences_triggered();
 
 private:
+	friend void SkinSwitchStorm::run(MainWindow& window);
+	void startSkinSwitchStorm();
 	void applySkinAndRebuild();
 	void executeStartAnalysis();
 	FilterTable* addTab(QString title, QString tooltip, QString configPath, QList<QString> lines);
